[Refactor] 사용자 관련 패키지 구조 변경
