+++ src/main/java/com/manchui/domain/user/dto/WrittenReviewInfo.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.dto.User;
+package com.manchui.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
