+++ src/main/java/com/manchui/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.entity;
+package com.manchui.domain.user.entity;
 
 import com.manchui.global.entity.Timestamped;
 import jakarta.persistence.Entity;
